% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{addrandomdiagnosis}
\alias{addrandomdiagnosis}
\title{Add Random Diagnosis}
\usage{
addrandomdiagnosis(all)
}
\arguments{
\item{all}{The data frame to add a Diagnosis column to}
}
\value{
The input data frame with a Diagnosis column added
}
\description{
This function will add a random diagnosis to every subject. It works by appending a Diagnosis column to the input data frame 'all'. Half of the diagnosis will be 'HC' (healthy control) and the other half will be 'AD' (Alzheimer's Disease).
}
\examples{
data <- generaterandomsubjects()
addrandomdiagnosis(data)
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
