% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{addrandomgender}
\alias{addrandomgender}
\title{Add Random Gender}
\usage{
addrandomgender(all)
}
\arguments{
\item{all}{The data frame to add a gender column to}
}
\value{
The input data frame with a Gender column added
}
\description{
This function will add a random gender to every subject. It works by appending a Gender column to the input data frame 'all'. Half of the diagnosis will be Male and the other half will be Female.
}
\examples{
data <- generaterandomsubjects()
addrandomgender(data)
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
