% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{adni.printfilelocations}
\alias{adni.printfilelocations}
\title{ADNI Print File Locations}
\usage{
adni.printfilelocations()
}
\description{
Prints the location of two files required to merge the (baseline) Diagnosis, Age, Gender and MMSE for ADNI subjects with the post-processed 'Freesurfer' data
}
\examples{
adni.setfiles("DXSUM_PDXCONV_ADNIALL.csv", "ADNIMERGE.csv")
adni.printfilelocations()
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
