% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{adni.setfiles}
\alias{adni.setfiles}
\title{ADNI Set Files}
\usage{
adni.setfiles(dxsumLocation, adnimergeLocation)
}
\arguments{
\item{dxsumLocation}{The filepath to DXSUM_PDXCONV_ADNIALL.csv}

\item{adnimergeLocation}{The filepath to ADNIMERGE.csv}
}
\description{
Points the location of two files required to merge the (baseline) Diagnosis, Age, Gender and MMSE for ADNI subjects with the post-processed 'Freesurfer' data
}
\examples{
adni.setfiles("DXSUM_PDXCONV_ADNIALL.csv", "ADNIMERGE.csv")
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
