% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{caddementia.printfilelocations}
\alias{caddementia.printfilelocations}
\title{CAD Dementia Print File Locations}
\usage{
caddementia.printfilelocations()
}
\description{
Prints the location of two files required to merge the Diagnosis, Age, and Gender for CAD Dementia subjects with the post-processed 'Freesurfer' data
}
\examples{
caddementia.setfiles("train.txt", "test.txt")
caddementia.printfilelocations()
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
