% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{caddementia.setfiles}
\alias{caddementia.setfiles}
\title{CAD Dementia Set Files}
\usage{
caddementia.setfiles(trainLocation, testLocation)
}
\arguments{
\item{trainLocation}{The filepath to train.txt}

\item{testLocation}{The filepath to test.txt
caddementia.setfiles("train.txt","test.txt")}
}
\description{
Points to the location of the two files required to merged the Diagnosis, Age and Gender for CAD Dementia subjects with the post-processed CAD Dementia data
}
\details{
Data can be accessed: \url{https://caddementia.grand-challenge.org/}
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
