% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{eliminateabnormalities.cols}
\alias{eliminateabnormalities.cols}
\title{Eliminate Abnormal Columns}
\usage{
eliminateabnormalities.cols(data, verbose = T)
}
\arguments{
\item{data}{The data frame to remove abnormal columns from}

\item{verbose}{Whether to print a log of what was removed}
}
\value{
The data frame with abnormal columns removed
}
\description{
Will remove columns from a data frame of data generated by 'Freesurfer' that are abnormal - columns where all values are zero
}
\examples{
data <- generaterandomsubjects()
eliminateabnormalities.cols(data)
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
