% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{eliminateabnormalities.rows}
\alias{eliminateabnormalities.rows}
\title{Eliminate Abnormal Rows}
\usage{
eliminateabnormalities.rows(data, verbose = T)
}
\arguments{
\item{data}{The data frame to remove abnormal rows from}

\item{verbose}{Whether to print a log of what was removed}
}
\value{
The data frame with abnormal rows removed
}
\description{
Will remove rows from a data frame of data generated by 'Freesurfer' that are abnormal - such as they have values of NA in them
}
\examples{
data <- generaterandomsubjects(10)
eliminateabnormalities.rows(data)
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
