% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{extract.brain.features}
\alias{extract.brain.features}
\title{Extract All MRI Brain Features}
\usage{
extract.brain.features(data, additionalFields = c())
}
\arguments{
\item{data}{The subject data to extract from}

\item{additionalFields}{Any additional fields to extract data from}
}
\value{
The MRI brain features
}
\description{
This command takes a data frame as input and extracts all the features segmented by 'Freesurfer'. Note that the MRIs must be processed with the 'hippocampal-subfields' flag when 'Freesurfer' is invoked.
}
\examples{
data <- generaterandomsubjects()
extract.brain.features(data)
}
\author{
Fabio Cappello
}
