% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{extract.byname}
\alias{extract.byname}
\title{Extract Features By Group Name}
\usage{
extract.byname(data, fieldGroupName, additionalFields = c())
}
\arguments{
\item{data}{Subject data to extract from}

\item{fieldGroupName}{The group field name to extract}

\item{additionalFields}{Any additional fields to extract}
}
\value{
The extracted fields
}
\description{
Extracts features in group specified by a string, groups that can be used are in normalise.listfieldgroups()
}
\examples{
data <- generaterandomsubjects()
extract.byname(data, "subcortical")
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
