% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{fsdirectorycheck}
\alias{fsdirectorycheck}
\title{'Freesurfer' Subject Directory Check}
\usage{
fsdirectorycheck(subjectDir, checkForHippocampalSubfieldsError = T)
}
\arguments{
\item{subjectDir}{The directory containing the subject subdirectories}

\item{checkForHippocampalSubfieldsError}{Whether to check to missing files which are to do with the "hippo-subfields" flag being called}
}
\value{
The number of errors found
}
\description{
Crawls through the subdirectory of subjects looking for missing file that would affect the fsimport() process
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
