% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{fsimport.listfields}
\alias{fsimport.listfields}
\title{'Freesurfer' Import (List Fields)}
\usage{
fsimport.listfields()
}
\description{
This function assumes all the subdirectories of subjectDir are subjects who have been processed in 'Freesurfer'. In then runs various 'Freesurfer' scripts to extract the specified fields into an R data frame.
}
\details{
Lists the field groups which can be imported from the subject are:
The specified fields can be:

lh.cortical.thickness = left hemisphere cortical thicknesses
rh.cortical.thickness = right hemisphere cortical thicknesses
lh.cortical.volume = left hemisphere cortical volumes
rh.cortical.volume = right hemisphere cortical volumes
lh.cortical.thickness.std = left hemisphere cortical thickness standard deviations
rh.cortical.thickness.std = right hemisphere cortical thickness standard deviations
lh.cortical.area = left hemisphere cortical surface areas
rh.cortical.area = right hemisphere cortical surface areas
lh.subcortical = left hemisphere subcortical volumes
rh.subcortical = right hemisphere subcortical volumes
lh.hippocampal = left hemisphere hippocampal volumes
rh.hippocampal = right hemisphere hippocampal volumes
}
\examples{
fsimport.listfields()
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
