% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{generaterandomsubjects}
\alias{generaterandomsubjects}
\title{Generate Random Subjects}
\usage{
generaterandomsubjects(num = 40)
}
\arguments{
\item{num}{The number of subjects to generate}
}
\value{
The generated subjects
}
\description{
Generate a data frame of random subjects whose fields match what would be imported from 'Freesurfer. This function is used for testing.
}
\examples{
generaterandomsubjects(num = 500)
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
