% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{get.hemisphere.side}
\alias{get.hemisphere.side}
\title{Get Hemisphere Side}
\usage{
get.hemisphere.side(name)
}
\arguments{
\item{name}{The name of the feature to return the hemisphere of}
}
\value{
The side of the hemisphere the feature belongs to ("left" or "right"). If it belongs to neither of these, "central" is returned
}
\description{
Given the name of a feature, will return a string as to whether it belongs to the left or the right hemisphere. If it belongs to neither, it is assumed that the feature is central
}
\examples{
get.hemisphere.side("Right.vessel")
get.hemisphere.side("lhCortexVol")
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
