% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{get.opposite.hemisphere.measurement}
\alias{get.opposite.hemisphere.measurement}
\title{Get Opposite Hemisphere Measurement}
\usage{
get.opposite.hemisphere.measurement(name, verbose = F, verbose_warn = F)
}
\arguments{
\item{name}{The name of the feature to return the corresponding measurement of the opposite hemisphere of}

\item{verbose}{If a corresponding feature doesn't exist, a message will be printed if this is true}

\item{verbose_warn}{If a corresponding feature doesn't exist, a message will be printed as a warning if this is true}
}
\value{
Name of the feature on the other hemisphere (or NULL if the feature does not exist)
}
\description{
Given a left hemisphere measurement, will return the equivalent measure on the right hemisphere. If there is no equivalent feature, NULL will be returned
}
\examples{
get.opposite.hemisphere.measurement("Right.vessel")
get.opposite.hemisphere.measurement("lhCortexVol")
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
