% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{getfieldgroup}
\alias{getfieldgroup}
\title{Get Field Group of Feature}
\usage{
getfieldgroup(name, method = 1)
}
\arguments{
\item{name}{Name of the feature}

\item{method}{The type of field groups that are returned
method = 1: {hippocampal, subcortical, corticalthicknessstds, corticalareas, corticalthicknesses, corticalvolumes}
method = 2: {volume, area, thickness, thicknessstd}}
}
\value{
The field group the name belongs to
}
\description{
Given the name of a feature, this function gets what type of field group it belongs to, i.e. subcortical volume
}
\examples{
getfieldgroup("left.CA1", method = 1)
getfieldgroup("left.CA1", method = 2)
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
