% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{getfshome}
\alias{getfshome}
\title{Get 'Freesurfer' Home}
\usage{
getfshome()
}
\value{
The directory 'Freesurfer' is installed
}
\description{
This command is used to get the base directory where 'Freesurfer' is installed as set using the command setfshome().
}
\examples{
setfshome("/Applications/freesurfer/")
getfshome()
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
