% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{getfsversion}
\alias{getfsversion}
\title{Get 'Freesurfer' Version}
\usage{
getfsversion(subjectDir)
}
\arguments{
\item{subjectDir}{The directory of the subject to get the 'Freesurfer' version of}
}
\value{
The version number
}
\description{
This command is used to get the version number of 'Freesurfer' which a certain subject was processed with
}
\examples{
\dontrun{
getfsversion("/Users/alex/Desktop/Subjects/002_S_0413/")
}
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
