% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{getnamesofareas}
\alias{getnamesofareas}
\title{Get Names of Areas}
\usage{
getnamesofareas(data = NULL, excludeFields = NULL)
}
\arguments{
\item{data}{Your subject data frame, if you have removed any columns from your data frame, the function will only return the areas in this data frame. If this parameter is NULL then a new data frame will be randomly generated}

\item{excludeFields}{A vector of areas names to exclude, to exclude nothing set this parameter to NULL}
}
\value{
A vector of the names of all the features which are areas
}
\description{
Gets the names of all the features that are areas
}
\examples{
getnamesofareas()
getnamesofareas(NULL, NULL)
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
