% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{getnamesofthicknesses}
\alias{getnamesofthicknesses}
\title{Get Names of Thicknesses}
\usage{
getnamesofthicknesses(data = NULL, excludeFields = NULL)
}
\arguments{
\item{data}{Your subject data frame, if you have removed any columns from your data frame, the function will only return the thicknesses in this data frame. If this parameter is NULL then a new data frame will be randomly generated}

\item{excludeFields}{A vector of thickness names to exclude, to exclude nothing set this parameter to NULL}
}
\value{
A vector of the names of all the features which are thicknesses
}
\description{
Gets the names of all the features that are thicknesses
}
\examples{
getnamesofthicknesses()
getnamesofthicknesses(NULL, NULL)
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
