% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{getnamesofvolumes}
\alias{getnamesofvolumes}
\title{Get Names of Volumes}
\usage{
getnamesofvolumes(data = NULL,
  excludeFields = c("EstimatedTotalIntraCranialVol"))
}
\arguments{
\item{data}{Your subject data frame, if you have removed any columns from your data frame, the function will only return the volumes in this data frame. If this parameter is NULL then a new data frame will be randomly generated}

\item{excludeFields}{A vector of volume names to exclude, to exclude nothing set this parameter to NULL}
}
\value{
A vector of the names of all the features which are volumes
}
\description{
Gets the names of all the features that are volumes
}
\examples{
getnamesofvolumes()
getnamesofvolumes(NULL, NULL)
getnamesofvolumes(NULL, excludeFields = c("left.CA1", "EstimatedTotalIntraCranialVol"))
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
