% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{is.cortical}
\alias{is.cortical}
\title{Is Feature Cortical}
\usage{
is.cortical(fieldName)
}
\arguments{
\item{fieldName}{The field name of the feature to check is a cortical measurement}
}
\value{
Whether the feature is cortical
}
\description{
Checks whether a feature is a cortical measurement
}
\examples{
is.cortical("lh.bankssts.thickness")
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
