% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{is.corticalarea}
\alias{is.corticalarea}
\title{Is Feature Cortical Area}
\usage{
is.corticalarea(fieldName)
}
\arguments{
\item{fieldName}{The field name of the feature to check is a cortical area}
}
\value{
Whether the feature is a cortical area
}
\description{
Checks whether a feature is a cortical area
}
\examples{
is.corticalarea("lh.bankssts.area")
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
