% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{is.corticalthickness}
\alias{is.corticalthickness}
\title{Is Feature Cortical Thickness}
\usage{
is.corticalthickness(fieldName)
}
\arguments{
\item{fieldName}{The field name of the feature to check is a cortical thickness}
}
\value{
Whether the feature is a cortical thickness
}
\description{
Checks whether a feature is a cortical thickness
}
\examples{
is.corticalthickness("lh.bankssts.thickness")
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
