% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{is.corticalthicknessstd}
\alias{is.corticalthicknessstd}
\title{Is Feature Cortical Thickness Standard Deviation}
\usage{
is.corticalthicknessstd(fieldName)
}
\arguments{
\item{fieldName}{The field name of the feature to check is a cortical thickness standard deviation}
}
\value{
Whether the feature is a cortical thickness standard deviation
}
\description{
Checks whether a feature is a cortical thickness standard deviation
}
\examples{
is.corticalthicknessstd("lh.bankssts.thicknessstd")
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
