% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{is.corticalvolume}
\alias{is.corticalvolume}
\title{Is Feature Cortical Volume}
\usage{
is.corticalvolume(fieldName)
}
\arguments{
\item{fieldName}{The field name of the feature to check is a cortical volume}
}
\value{
Whether the feature is a cortical volume
}
\description{
Checks whether a feature is a cortical volume
}
\examples{
is.corticalvolume("lh.bankssts.volume")
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
