% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{is.hippocampalvolume}
\alias{is.hippocampalvolume}
\title{Is Feature Hippocampal Volume}
\usage{
is.hippocampalvolume(fieldName)
}
\arguments{
\item{fieldName}{The field name of the feature to check is a hippocampal volume}
}
\value{
Whether the feature is a hippocampal volume
}
\description{
Checks whether a feature is a hippocampal volume
}
\examples{
is.hippocampalvolume("right.fimbria")
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
