% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{is.subcorticalvolume}
\alias{is.subcorticalvolume}
\title{Is Feature Subcortical Volume}
\usage{
is.subcorticalvolume(fieldName)
}
\arguments{
\item{fieldName}{The field name of the feature to check is a subcortical volume}
}
\value{
Whether the feature is a subcortical volume
}
\description{
Checks whether a feature is a subcortical volume
}
\examples{
is.subcorticalvolume("Brain.Stem")
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
