% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{isfsfeature}
\alias{isfsfeature}
\title{Is Feature Created by 'Freesurfer'}
\usage{
isfsfeature(name)
}
\arguments{
\item{name}{Name of the feature}
}
\value{
Whether the feature was created by 'Freesurfer'
}
\description{
Given the name of a feature, this function returns whether it was a feature generated the by 'Freesurfer' processing stream
}
\examples{
isfsfeature("left.CA1")
isfsfeature("Age")
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
