% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{ixi.mergewithfreesurferoutput}
\alias{ixi.mergewithfreesurferoutput}
\title{'IXI' Merge Study Information With 'Freesurfer' Output}
\usage{
ixi.mergewithfreesurferoutput(ixi.mri.data, verbose = T)
}
\arguments{
\item{ixi.mri.data}{The data frame containing the data generated by 'Freesurfer' of the 'IXI' subjects}

\item{verbose}{Whether to log information to the console}
}
\description{
Merges the external 'IXI' study data (specified with 'IXI' Set Files) with the data generated by 'Freesurfer'
}
\details{
Data can be accessed: \url{http://brain-development.org/ixi-dataset/}

Requires a modification of IXI.xls
}
\author{
Fabio Cappello, Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
