% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{ixi.setfile}
\alias{ixi.setfile}
\title{IXI Set File}
\usage{
ixi.setfile(location)
}
\arguments{
\item{location}{The filepath to IXI.csv
ixi.setfile("IXI.csv")}
}
\description{
Points to the location of the two files required to merge the Diagnosis, Age and Gender for IXI subjects with the post-processed IXI data
}
\details{
Data can be accessed: \url{http://brain-development.org/ixi-dataset/}

Requires a modification of IXI.xls
}
\author{
Fabio Cappello
}
