% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{normalise.listfieldgroups}
\alias{normalise.listfieldgroups}
\title{List Normalisation Field Groups}
\usage{
normalise.listfieldgroups()
}
\value{
A list of the normalisation field groups
}
\description{
Lists all the available field groups that the normalisation can operate on
}
\examples{
normalise.listfieldgroups()
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
