% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{normalise.listmethods}
\alias{normalise.listmethods}
\title{List Normalisation Methods}
\usage{
normalise.listmethods()
}
\value{
A list of the normalisation methods
}
\description{
Lists all the available normalisation methods
}
\examples{
normalise.listmethods()
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
