% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\docType{package}
\name{rsurfer-package}
\alias{rsurfer}
\alias{rsurfer-package}
\title{RSurfer}
\description{
A package for interfacing between R and 'Freesurfer'
}
\details{
This package provides functionality for importing MRI data which has been processed with 'Freesurfer' into R. It also provides functions for manipulating this data within R such as functionality to perform intracranial volume normalisation and field manipulation.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/AlexDiru/rsurfer/issues}
}

}
\author{
\strong{Maintainer}: Alexander Luke Spedding \email{alexspedding271@gmail.com}

Authors:
\itemize{
  \item Fabio Cappello
  \item Giuseppe di Fatta
}

}
