% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{searchforabnormalities.cols}
\alias{searchforabnormalities.cols}
\title{Search For Abnormalities (Columns)}
\usage{
searchforabnormalities.cols(data, verbose = T)
}
\arguments{
\item{data}{The data frame imported using fsimport}

\item{verbose}{Whether the print debug information}
}
\value{
Indices of abnormal columns
}
\description{
Looks for columns which have been exported from 'Freesurfer' and may cause classification problems, for example, my personal abnormal columns are Left.WM.hypointensities, Right.WM.hypointensities, Left.non.WM.hypointensities, Right.WM.hypointensities only have values of zero
}
\examples{
data <- generaterandomsubjects()
searchforabnormalities.cols(data)
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
