% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{searchforabnormalities.rows}
\alias{searchforabnormalities.rows}
\title{Search For Abnormalities (Rows)}
\usage{
searchforabnormalities.rows(data, verbose = T)
}
\arguments{
\item{data}{The data frame imported using fsimport}

\item{verbose}{Whether the print debug information}
}
\value{
Indices of abnormal rows
}
\description{
Looks for rows (subjects) which have been exported from 'Freesurfer' and may cause classification problems, for example, rows with NAs in
}
\examples{
data <- generaterandomsubjects()
searchforabnormalities.rows(data)
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
