% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{setfshome}
\alias{setfshome}
\title{Set 'Freesurfer' Home}
\usage{
setfshome(freesurferDirectory)
}
\arguments{
\item{freesurferDirectory}{The directory 'Freesurfer' is installed to}
}
\value{
None
}
\description{
This command is used to set the base directory where 'Freesurfer' is installed. My installation of 'Freesurfer' is located in: "/Applications/freesurfer/"; thus for my installation location setfshome("/Applications/freesurfer/") would be used.
}
\examples{
setfshome("/Applications/freesurfer")
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
