% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{test.fieldextraction}
\alias{test.fieldextraction}
\title{Test Field Extraction}
\usage{
test.fieldextraction(subjectTestDir)
}
\arguments{
\item{subjectTestDir}{The directory containing the subject subdirectories}
}
\description{
Tests the field extraction functions are working correctly on imported data from subjectDir. Note: requires the subjects to be processed with the "hippocampal-subfields" flag.
}
\examples{
\dontrun{
test.fieldextraction("/Users/alex/Desktop/Subjects")
}
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
