% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{test.importing}
\alias{test.importing}
\title{Test Importing}
\usage{
test.importing(subjectTestDir)
}
\arguments{
\item{subjectTestDir}{The directory containing the subject subdirectories}
}
\description{
Calls fsimport() with different parameters on the subjects in subjectDir to test it is working correctly. Note: requires the subjects to be processed with the "hippocampal-subfields" flag.
}
\examples{
\dontrun{
test.importing("/Users/alex/Desktop/Subjects")
}
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
