% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsurfer.R
\name{test.normalisation}
\alias{test.normalisation}
\title{Test Normalisation}
\usage{
test.normalisation()
}
\description{
Tests the normalisation functions are working correctly on randomly generated data.
}
\examples{
\dontrun{
test.normalisation()
}
}
\author{
Alexander Luke Spedding, \email{alexspedding271@gmail.com}
}
